tidy up structure. sort of. need to think about this
